requirements.txt
fastapi
uvicorn[standard]
pillow
tensorflow
python-multipart
